separate code: DOM interaction and data management
